requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.3
fastjsonschema==2.21.2
Flask==3.0.3
fonttools==4.59.2
google-auth==2.41.1
google-genai==1.45.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupytext==1.17.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.4.0
plotly==6.1.0
pyadi-iio==0.0.19
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pylibiio==0.25
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
rsa==4.9.1
SigMF==1.2.10
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.0.6
zipp==3.23.0
